server/requirements.txt
asgiref==3.7.2
azure-core==1.29.5
azure-identity==1.15.0
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.7
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
isodate==0.6.1
msal==1.25.0
msal-extensions==1.0.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
